requirements.txt
streamlit
google-genai
pinecone
PyMuPDF
python-dotenv
